pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mmetl"
version = "0.0.1"
description = "Minimal mmETL pipeline framework"
readme = "README.md"
authors = [{ name = "mmETL Team" }]
requires-python = ">=3.11"
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "typer[all]",
  "pydantic>=2",
  "pyyaml",
  "redis>=5.0",
  "rich>=13.0",
  "duckdb>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "mypy>=1.8",
  "ruff>=0.5.0",
  "black>=24.3.0",
]

[project.scripts]
mmetl = "mmETL.cli.main:app"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
